ci(ssl): 👷 improve ssl generator checker

+++ ssl.sh
@@ -1,12 +1,18 @@
 export DOMAIN=$1
+# cspell:ignore fullchain letsencrypt
 
-if ! command -v certbot &>/dev/null; then
-  sudo snap install core
-  sudo snap refresh core
-  sudo snap install --classic certbot
-  sudo ln -s /snap/bin/certbot /usr/bin/certbot
-  sudo snap set certbot trust-plugin-with-root=ok
-  sudo snap install certbot-dns-digitalocean
+FILE=/etc/letsencrypt/live/$DOMAIN/fullchain.pem
+
+if ! test -f "$FILE"; then
+
+  if ! command -v certbot &>/dev/null; then
+    sudo snap install core
+    sudo snap refresh core
+    sudo snap install --classic certbot
+    sudo ln -s /snap/bin/certbot /usr/bin/certbot
+    sudo snap set certbot trust-plugin-with-root=ok
+    sudo snap install certbot-dns-digitalocean
+  fi
 
   # cspell:ignore certbot certonly
   certbot certonly \
@@ -18,4 +24,5 @@ if ! command -v certbot &>/dev/null; then
     --agree-tos \
     --non-interactive
   exit
+
 fi
